refactor to shorten each file
