+++ gluoncv/model_zoo/rcnn/mask_rcnn/__init__.py
@@ -3,3 +3,4 @@
 from __future__ import absolute_import
 
 from .mask_rcnn import *
+from .predefined_models import *
